Support for static members (#6)
This patch adds support to monitor when a class is loaded
or the static methods are executed. Then, for example, if a
class is loaded dynamically by spring when the test starts,
this is instrumented.

If fields are accessed and they do not have a related behavior
to the class, they should be ignored.